// src/contexts/Farms/types.ts
import { Contract } from 'web3-eth-contract'

export interface Farm {
  pid: number
  name: string
  type: string
  lpToken: string
  lpTokenAddress: string
  lpContract: Contract
  tokenAddress: string
  earnToken: string
  earnTokenAddress: string
  icon: React.ReactNode
  id: string
  tokenSymbol: string
  buylink : string
  lplink : string
}

export interface FarmsContext {
  farms: Farm[]
  // sales: Sale[]
  unharvested: number
}